src/services/dashboard/pyproject.toml
[project]
name = "gladys-dashboard"
version = "0.1.0"
description = "GLADyS Dashboard V2"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "python-multipart",
    "uvicorn",
    "sse-starlette",
    "jinja2",
    "structlog",
    "grpcio",
    "grpcio-tools",
    "psycopg2-binary",
]

[project.optional-dependencies]
test = ["pytest", "httpx", "pytest-playwright", "pytest-asyncio"]

[tool.setuptools]
packages = ["backend"]

[tool.uv]
package = true
